fixed backend and interview frontend

+++ backend/interviewer-service/src/main/java/com/hackerrank/interviewer_service/Repository/AllRequirementRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface AllRequirementRepository extends JpaRepository<AllRequirements, Integer> {
+public interface AllRequirementRepository extends JpaRepository<AllRequirements, Long> {
     Optional<AllRequirements> findByRequirementName(String requirementName);
 }
